# openapi.yaml
openapi: 3.0.1
info:
  title: kaptain
  description: The kaptain API
  version: "0.0.1"
servers:
- url: /v1/
paths:
  /{repo}/{chart}:
    get:
      summary: Get chart values
      operationId: app.get_repo_chart
      parameters:
        - in: path
          name: repo
          schema:
            type: string
          required: true
          description: Repo name
        - in: path
          name: chart
          schema:
            type: string
          required: true
          description: Chart name
      responses:
        200:
          description: Return chart values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChartValues'
        404:
          description: Repo or chart doesn't exist
        422:
          description: Repo files couldn't be processed
    put:
      summary: Put chart values
      operationId: app.put_repo_chart
      parameters:
        - in: path
          name: repo
          schema:
            type: string
          required: true
          description: Repo name
        - in: path
          name: chart
          schema:
            type: string
          required: true
          description: Chart name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChartValues'
      responses:
        200:
          description: Values updated
        404:
          description: Repo or chart doesn't exist
        422:
          description: Repo files couldn't be processed
components:
  schemas:
    ChartValues:
      required:
      - values
      type: object
      properties:
        values:
          type: array
          items:
            $ref: '#/components/schemas/ChartValue'
    ChartValue:
      required:
      - path
      - value
      properties:
        path:
          type: string
          minLength: 1
          example: "server.image.tag"
        value:
          oneOf:
          - type: string
          - type: number
          - type: integer
          example: "latest"